refactor(studio-ui): move components into per-component folders
Each components/*.tsx becomes ComponentName/index.tsx, with
props types and RowGrid's primaryKeyColumn helper split into
types/ and library/ — matching the hooks/useStudioConnection
convention. Tests move alongside their component. App.tsx now
imports from the components/hooks barrels.
