+++ packages/salve-db-studio/ui/src/components/InsertForm/__tests__/InsertForm.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
-import { InsertForm } from '../InsertForm';
-import type { IColumnInfo } from '../../types';
+import { InsertForm } from '../index';
+import type { IColumnInfo } from '../../../types';
 
 const columns: IColumnInfo[] = [
   { cid: 0, name: 'id', type: 'INTEGER', notnull: 0, dflt_value: null, pk: 1 },
