+++ packages/salve-db-studio/ui/src/components/InsertForm/types/IInsertFormProps.ts
@@ -1,0 +1,6 @@
+import type { IColumnInfo } from '../../../types';
+
+export interface IInsertFormProps {
+  columns: IColumnInfo[];
+  onSubmit: (values: Record<string, string>) => void;
+}
