+++ packages/salve-db-studio/ui/src/components/RowGrid/library/index.ts
@@ -1,0 +1,1 @@
+export * from './primaryKeyColumn';
