+++ packages/salve-db-studio/ui/src/components/RowGrid/library/primaryKeyColumn.ts
@@ -1,0 +1,5 @@
+import type { IColumnInfo } from '../../../types';
+
+export function primaryKeyColumn(columns: IColumnInfo[]): string | null {
+  return columns.find((c) => Number(c.pk) > 0)?.name ?? null;
+}
