+++ packages/salve-db-studio/ui/src/components/SqlRunner/types/IResult.ts
@@ -1,0 +1,4 @@
+export interface IResult {
+  text: string;
+  isError: boolean;
+}
