+++ packages/salve-db-studio/ui/src/components/SqlRunner/types/ISqlRunnerProps.ts
@@ -1,0 +1,5 @@
+import type { Row } from '../../../types';
+
+export interface ISqlRunnerProps {
+  runSql: (sql: string) => Promise<Row[]>;
+}
