+++ packages/salve-db-studio/ui/src/components/SqlRunner/types/index.ts
@@ -1,0 +1,2 @@
+export type * from './ISqlRunnerProps';
+export type * from './IResult';
