+++ packages/salve-db-studio/ui/src/components/StatusBadge/types/IStatusBadgeProps.ts
@@ -1,0 +1,3 @@
+export interface IStatusBadgeProps {
+  connected: boolean;
+}
