+++ packages/salve-db-studio/ui/src/components/StatusBadge/types/index.ts
@@ -1,0 +1,1 @@
+export type * from './IStatusBadgeProps';
