+++ packages/salve-db-studio/ui/src/components/TableList/types/ITableListProps.ts
@@ -1,0 +1,5 @@
+export interface ITableListProps {
+  tables: string[];
+  currentTable: string | null;
+  onSelect: (name: string) => void;
+}
