+++ packages/salve-db-studio/ui/src/components/Toast/types/IToastProps.ts
@@ -1,0 +1,4 @@
+export interface IToastProps {
+  message: string | null;
+  onDismiss: () => void;
+}
