+++ packages/salve-db-studio/ui/src/components/Toast/types/index.ts
@@ -1,0 +1,1 @@
+export type * from './IToastProps';
